+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- `flip <task(s)>` - flips tasks as done/undone
+
 ## [0.2.0] - 2026-08-24
 
 - `add <text(s)>` - add 1 or more tasks